+++ src/graphics/graphics_module.cpp
@@ -5,8 +5,6 @@
 
 namespace gecko::graphics {
 
-static GraphicsModule s_GraphicsModule;
-
 constexpr ::gecko::Label GraphicsModule::RootLabel() const noexcept
 {
   return labels::Graphics;
@@ -21,17 +19,6 @@ bool GraphicsModule::Startup(::gecko::IModuleRegistry& /*modules*/) noexcept
 void GraphicsModule::Shutdown(::gecko::IModuleRegistry& /*modules*/) noexcept
 {
   GECKO_FUNC(labels::Graphics);
-}
-
-::gecko::ModuleRegistration InstallGraphicsModule(
-    ::gecko::IModuleRegistry& modules) noexcept
-{
-  return modules.RegisterStatic(s_GraphicsModule);
-}
-
-::gecko::IModule& GetModule() noexcept
-{
-  return s_GraphicsModule;
 }
 
 }  // namespace gecko::graphics