+++ template/module/src/__MODULE__/__MODULE___module.cpp
@@ -6,8 +6,6 @@
 
 namespace gecko::__MODULE__ {
 
-static __MODULE_CAMEL__Module s___MODULE_CAMEL__Module;
-
 constexpr ::gecko::Label __MODULE_CAMEL__Module::RootLabel() const noexcept
 {
   return labels::__MODULE_CAMEL__;
@@ -23,17 +21,6 @@ void __MODULE_CAMEL__Module::Shutdown(
     ::gecko::IModuleRegistry& modules) noexcept
 {
   GECKO_FUNC(labels::__MODULE_CAMEL__);
-}
-
-::gecko::ModuleRegistration Install__MODULE_CAMEL__Module(
-    ::gecko::IModuleRegistry& modules) noexcept
-{
-  return modules.RegisterStatic(s___MODULE_CAMEL__Module);
-}
-
-::gecko::IModule& GetModule() noexcept
-{
-  return s___MODULE_CAMEL__Module;
 }
 
 }  // namespace gecko::__MODULE__